automllab/requirements.txt
anthropic==0.36.0
sentence-transformers==3.3.0
scikit-learn==1.5.0
numpy==1.26.0
pandas==2.2.0
torch==2.2.0
torchmetrics==1.4.0
gitpython==3.1.43
fastapi==0.115.0
uvicorn[standard]==0.30.0
reportlab==4.2.5
pydantic>=2.0.0
pytest==8.3.0
pytest-asyncio==0.24.0
